refactor: streamline imports in engine modules
